Cleaned up the example to fit new style guide
