+++ examples/posix_demo/README.md
@@ -1,0 +1,11 @@
+To try this example, run the following from the top-level directory (OBC-firmware):
+
+```bash
+make -f examples/posix_demo/Makefile # Build
+./examples/posix_demo/build/posix_example # Run
+```
+
+To remove the build files, run the following:
+```bash
+make -f examples/posix_demo/Makefile clean
+```
