+++ examples/posix_demo/include/posix_example.h
@@ -1,0 +1,6 @@
+#ifndef EXAMPLES_POSIX_DEMO_INCLUDE_POSIX_EXAMPLE_H_
+#define EXAMPLES_POSIX_DEMO_INCLUDE_POSIX_EXAMPLE_H_
+
+void vDummyTask(void * pvParameters);
+
+#endif /* EXAMPLES_POSIX_DEMO_INCLUDE_POSIX_EXAMPLE_H_ */
