--- examples/posix_makefile_example/README.md
@@ -1,11 +1,0 @@
-To try this example, run the following from the top-level directory (OBC-firmware):
-
-```bash
-make -f examples/posix_makefile_example/Makefile # Build
-./examples/posix_makefile_example/build/posix_example # Run
-```
-
-To remove the build files, run the following:
-```bash
-make -f examples/posix_makefile_example/Makefile clean
-```
