--- examples/posix_makefile_example/include/posix_example.h
@@ -1,15 +1,0 @@
-#ifndef POSIX_EXAMPLE_H
-#define POSIX_EXAMPLE_H
-
-/**
- * @brief Add two numbers
- * @param a 
- * @param b 
- * 
- * @return The sum of a and b
- */
-int add(int a, int b);
-
-void vDummyTask(void * pvParameters);
-
-#endif
