+++ posix/include/console.h
@@ -1,16 +1,22 @@
-#ifndef CONSOLE_H
-    #define CONSOLE_H
+#ifndef POSIX_INCLUDE_CONSOLE_H_
+#define POSIX_INCLUDE_CONSOLE_H_
 
-    #ifdef __cplusplus
-        extern "C" {
-    #endif
-
-    void console_init( void );
-    void console_print( const char * fmt,
-                        ... );
-
-    #ifdef __cplusplus
-        }
-    #endif
-
+#ifdef __cplusplus
+    extern "C" {
 #endif
+
+/*
+ * Initialize the mutex for the console.
+ */
+void initConsole(void);
+
+/*
+ * Print a string to the console.
+ */
+void printConsole(const char * fmt, ...);
+
+#ifdef __cplusplus
+    }
+#endif
+
+#endif /* POSIX_INCLUDE_CONSOLE_H_ */
