+++ posix/source/console.c
@@ -6,15 +6,15 @@
 
 SemaphoreHandle_t xStdioMutex;
 
-void console_init( void ) {
+void initConsole(void) {
     xStdioMutex = xSemaphoreCreateMutex();
 }
 
-void console_print( const char * fmt, ... ) {
+void printConsole(const char * fmt, ...) {
     va_list vargs;
-    va_start( vargs, fmt );
-    xSemaphoreTake( xStdioMutex, portMAX_DELAY );
-    vprintf( fmt, vargs );
-    xSemaphoreGive( xStdioMutex );
-    va_end( vargs );
+    va_start(vargs, fmt);
+    xSemaphoreTake(xStdioMutex, portMAX_DELAY);
+    vprintf(fmt, vargs);
+    xSemaphoreGive(xStdioMutex);
+    va_end(vargs);
 }
